1.42 - Feature Additions / Changes
Added a searchbar and search button to the game tab.

Modified the colour scheme to fit the wireframe.
Modified the title, menu tab and misc tab labels to uppercase to fit the wireframe.
Modified the colour of each menu tab to better fit the colour scheme.
Modified the code definitions for button hovering, menu tabs and quit confirmation.
Modified the UI's colour variables to mitigate hardcoding issues.